MapApp Android Application
I created this Android Application as my course project and also to get a handson experience with Android and google api's. As its name its an app to find nearby places such as nearby schools, hospitals, restaurants. Also to get directions from a particular place. This App is an attempt to club a texting app and google maps app together. The code for the texting app is yet to be written. But the app can be used as a google map with an additional feature of nearby common places.